Menu refactor to fix circular dependency
