+++ packages/react-tinacms-editor/src/state/plugins/Menu/Image/FloatingImageMenu.tsx
