+++ packages/react-tinacms-editor/src/state/plugins/Menu/Menu.tsx
@@ -18,12 +18,11 @@ limitations under the License.
 
 import { EditorView } from 'prosemirror-view'
 import * as React from 'react'
-import { useState, useRef, useEffect, useLayoutEffect, useMemo } from 'react'
+import { useState, useRef, useEffect, useLayoutEffect } from 'react'
 
 import { markControl } from './markControl'
 import { FormattingDropdown } from './FormattingDropdown'
 import { FloatingTableMenu } from './FloatingTableMenu'
-import ImageMenu from './ImageMenu'
 import {
   toggleBulletList,
   toggleOrderedList,
@@ -33,7 +32,7 @@ import { imagePluginKey } from '../Image'
 import { wrapIn, setBlockType } from 'prosemirror-commands'
 import { EditorState } from 'prosemirror-state'
 import { findParentNodeOfType } from 'prosemirror-utils'
-import styled, { css, ThemeProvider } from 'styled-components'
+import { ThemeProvider } from 'styled-components'
 import {
   BoldIcon,
   CodeIcon,
@@ -45,10 +44,16 @@ import {
   UnderlineIcon,
   UnorderedListIcon,
 } from '@tinacms/icons'
-import { radius, color, padding } from '@tinacms/styles'
 import { UndoControl, RedoControl } from './historyControl'
-import { MenuPortalProvider, useMenuPortal } from './MenuPortal'
+import { MenuPortalProvider } from './MenuPortal'
 import { FloatingLinkForm } from '../links/FloatingLinkForm'
+import FloatingImageMenu from './Image/FloatingImageMenu'
+import {
+  MenuButton,
+  MenuPlaceholder,
+  MenuWrapper,
+  MenuContainer,
+} from './MenuButton'
 
 interface Props {
   bottom?: boolean
@@ -180,7 +185,6 @@ export const Menu = (props: Props) => {
               <ItalicControl view={view} />
               <UnderlineControl view={view} />
               <LinkControl view={view} />
-              {/* <ImageControl view={view} bottom={bottom} /> */}
               {supportBlocks && <TableControl view={view} bottom={bottom} />}
               {supportBlocks && <QuoteControl view={view} bottom={bottom} />}
               {supportBlocks && <CodeControl view={view} bottom={bottom} />}
@@ -192,7 +196,7 @@ export const Menu = (props: Props) => {
           </MenuPortalProvider>
         </MenuWrapper>
         <FloatingTableMenu editorView={editorView} />
-        <ImageMenu editorView={editorView} />
+        <FloatingImageMenu editorView={editorView} />
         <FloatingLinkForm editorView={editorView} />
       </>
     </ThemeProvider>
@@ -285,196 +289,3 @@ const OrderedList = commandContrl(
   'Ordered List',
   'Ordered List'
 )
-
-type MenuPlaceholderProps = {
-  menuBoundingBox: any
-}
-
-const MenuPlaceholder = styled.div<MenuPlaceholderProps>`
-  color: transparent;
-  background: transparent;
-  pointer-events: none;
-  position: relative;
-  display: block;
-  height: ${props => props.menuBoundingBox.height}px;
-  width: ${props => props.menuBoundingBox.width}px;
-`
-
-type MenuWrapperProps = {
-  menuFixed: boolean
-  menuBoundingBox: any
-  menuFixedTopOffset: string
-}
-
-const MenuWrapper = styled.div<MenuWrapperProps>`
-  position: relative;
-  margin-bottom: 14px;
-  z-index: 900;
-
-  ${props =>
-    props.menuFixed &&
-    css`
-      position: fixed;
-      width: ${props.menuBoundingBox.width}px;
-      top: ${props.menuFixedTopOffset};
-    `};
-`
-
-const MenuContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  position: relative;
-  top: 0;
-  width: 100%;
-  background-color: white;
-  border-radius: ${radius()};
-  box-shadow: 0px 2px 3px rgba(0, 0, 0, 0.12);
-  border: 1px solid ${color.grey(2)};
-  overflow: hidden;
-  z-index: 100;
-`
-
-const MenuItem = css`
-  flex: 1 1 24px;
-`
-
-export const MenuButton = styled.button<{
-  active?: boolean
-  disabled?: boolean
-  bottom?: boolean
-  ref?: any
-}>`
-  ${MenuItem}
-  background-color: ${p =>
-    p.active ? 'rgba(53, 50, 50, 0.05)' : 'transparent'};
-  color: ${p => (p.active ? '#0084ff' : color.grey(8))};
-  fill: ${p => (p.active ? '#0084ff' : color.grey(8))};
-  border: 1px solid ${color.grey(2)};
-  margin: -1px;
-  outline: none;
-  padding: 6px 4px;
-  transition: all 85ms ease-out;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  &:hover {
-    background-color: rgba(53, 50, 50, 0.09);
-  }
-  &:active {
-    color: #0084ff;
-    fill: #0084ff;
-    background-color: rgba(53, 50, 50, 0.05);
-  }
-  svg {
-    width: 20px;
-    height: 20px;
-  }
-  ${props =>
-    props.active &&
-    css`
-      color: #0084ff;
-      fill: #0084ff;
-      background-color: rgba(53, 50, 50, 0.05);
-    `};
-  ${props =>
-    props.disabled &&
-    css`
-      pointer-events: none;
-      color: #d1d1d1;
-      fill: #d1d1d1;
-    `};
-`
-
-export const MenuDropdownWrapper = styled.div`
-  ${MenuItem}
-  position: relative;
-
-  ${MenuButton} {
-    width: 100%;
-  }
-`
-
-export const MenuButtonDropdown = styled(
-  ({ children, open, triggerRef, ...styleProps }) => {
-    const MenuPortal = useMenuPortal()
-    const menuPortalRef = React.useRef<HTMLDivElement | null>(null)
-
-    const menuOffset = useMemo(() => {
-      if (!triggerRef.current || !menuPortalRef.current) return 0
-      const menuDropdownBoundingBox = triggerRef.current.getBoundingClientRect()
-      const menuPortalBoundingBox = menuPortalRef.current.getBoundingClientRect()
-      return menuDropdownBoundingBox.x - menuPortalBoundingBox.x
-    }, [triggerRef.current, menuPortalRef.current])
-
-    return (
-      <MenuPortal>
-        <Offset offset={menuOffset}>
-          <div ref={menuPortalRef} {...styleProps}>
-            {children}
-          </div>
-        </Offset>
-      </MenuPortal>
-    )
-  }
-)`
-  border-radius: ${radius()};
-  border: 1px solid #efefef;
-  display: block;
-  position: absolute;
-  bottom: -4px;
-  left: 0;
-  transform: translate3d(0, 100%, 0) scale3d(0.5, 0.5, 1);
-  opacity: 0;
-  pointer-events: none;
-  transition: all 85ms ease-out;
-  transform-origin: 0 0;
-  box-shadow: 0px 2px 3px rgba(0, 0, 0, 0.12), 0px 4px 8px rgba(48, 48, 48, 0.1);
-  background-color: white;
-  overflow: hidden;
-
-  ${props =>
-    props.open &&
-    css`
-      opacity: 1;
-      pointer-events: all;
-      transform: translate3d(0, 100%, 0) scale3d(1, 1, 1);
-    `};
-`
-
-const Offset = styled.div<{ offset: number }>`
-  position: absolute;
-  left: ${props => props.offset}px;
-`
-
-export const MenuOption = styled.div<{ disabled: boolean; active: boolean }>`
-  display: block;
-  padding: 8px 16px;
-  transition: all 85ms ease-out;
-  cursor: pointer;
-  &:first-child {
-    padding-top: ${padding('small')};
-  }
-  &:last-child {
-    padding-bottom: ${padding('small')};
-  }
-  &:hover {
-    background-color: ${color.grey(1)};
-    color: ${color.primary()};
-  }
-  &:active {
-    color: ${color.primary()};
-    fill: ${color.primary()};
-    background-color: rgba(53, 50, 50, 0.05);
-  }
-  ${props =>
-    props.active &&
-    css`
-      color: #0084ff;
-      fill: #0084ff;
-      background-color: rgba(53, 50, 50, 0.05);
-    `};
-`
