+++ packages/react-tinacms-editor/src/state/plugins/Menu/historyControl.tsx
@@ -22,7 +22,7 @@ import { RedoIcon, UndoIcon } from '@tinacms/icons'
 import { EditorView } from 'prosemirror-view'
 import { undo, redo, undoDepth, redoDepth } from 'prosemirror-history'
 
-import { MenuButton } from './Menu'
+import { MenuButton } from './MenuButton'
 
 interface HistoryProps {
   view: EditorView
